implement logout + some movement

+++ worldserver/data/object/location.go
@@ -4,13 +4,13 @@ import "math"
 
 // Location represents an objects location within the world.
 type Location struct {
-	X, Y, Z, O float64
+	X, Y, Z, O float32
 }
 
 // Distance calculates the distance between two locations.
 func (loc *Location) Distance(other *Location) float64 {
 	return math.Sqrt(
-		math.Pow(loc.X-other.X, 2) +
-			math.Pow(loc.X-other.X, 2) +
-			math.Pow(loc.X-other.X, 2))
+		math.Pow(float64(loc.X-other.X), 2) +
+			math.Pow(float64(loc.X-other.X), 2) +
+			math.Pow(float64(loc.X-other.X), 2))
 }
